# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateUserRequest:
    properties:
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      photo_url:
        type: string
      role:
        enum:
        - admin
        - peserta
        type: string
      username:
        type: string
    required:
    - full_name
    - password
    - role
    - username
    type: object
...
  domain.UpdateUserPasswordRequest:
    properties:
      password:
        minLength: 6
        type: string
    required:
    - password
    type: object
  domain.UpdateUserRequest:
    properties:
      full_name:
        type: string
      photo_url:
        type: string
      role:
        enum:
        - admin
        - peserta
        type: string
      username:
        type: string
    required:
    - full_name
    - role
    - username
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      photo_url:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      message:
        example: Terjadi kesalahan
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Dev Team Pramuka CAT
    url: https://github.com/odealidj/pramuka-CAT
  description: REST API untuk platform Computer Assisted Test (CAT) Pramuka.
  termsOfService: http://swagger.io/terms/
  title: Pramuka CAT API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: Mengambil daftar seluruh pengguna dengan paginasi
      parameters:
      - description: 'Halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah per halaman (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.User'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Semua User
      tags:
      - Admin - User
    post:
      consumes:
      - application/json
      description: Admin membuat akun peserta atau admin baru
      parameters:
      - description: Data User Baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Bad Request
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat User Baru
      tags:
      - Admin - User
  /admin/users/{id}:
    delete:
      description: Admin menonaktifkan akun pengguna (data tidak benar-benar dihapus)
      parameters:
      - description: UUID User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Not Found
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus User (Soft-Delete)
      tags:
      - Admin - User
    get:
      description: Mengambil data detail satu pengguna berdasarkan ID
      parameters:
      - description: UUID User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "404":
          description: Not Found
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detail User
      tags:
      - Admin - User
    put:
      consumes:
      - application/json
      description: Admin memperbarui data profil pengguna
      parameters:
      - description: UUID User
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Bad Request
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Data User
      tags:
      - Admin - User
  /admin/users/{id}/password:
    put:
      consumes:
      - application/json
      description: Admin mengubah kata sandi pengguna
      parameters:
      - description: UUID User
        in: path
        name: id
        required: true
        type: string
      - description: Password Baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Password User
      tags:
      - Admin - User
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"